% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezirke.R
\name{zh_bezirk}
\alias{zh_bezirk}
\title{Get a single district by code}
\usage{
zh_bezirk(bezirk_code)
}
\arguments{
\item{bezirk_code}{Integer or character. The district code.}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble()}} with district details.
}
\description{
Get a single district by code
}
\examples{
\dontrun{
zh_bezirk(101)  # Affoltern
}
}
