% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezirke.R
\name{zh_bezirke}
\alias{zh_bezirke}
\title{List all districts (Bezirke)}
\usage{
zh_bezirke()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\code{gebietstyp_code}, \code{bezirk_code}, \code{bezirk_name}.
}
\description{
Retrieves all districts in the Canton of Zurich.
}
\examples{
\dontrun{
zh_bezirke()
}
}
