% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezirke.R
\name{zh_bezirke_suche}
\alias{zh_bezirke_suche}
\title{Search districts by name}
\usage{
zh_bezirke_suche(bezirkname, .include_query = FALSE)
}
\arguments{
\item{bezirkname}{Character scalar. (Partial) name to search for.}

\item{.include_query}{Logical. If \code{TRUE}, a \code{query} column with the
search term is added. Default \code{FALSE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{gemeinde_code},
\code{gemeinde_name}, \code{bezirk_code}, \code{bezirk_name},
\code{raumplanungsregion_code}, \code{raumplanungsregion_name}
(and optionally \code{query}).
}
\description{
Search districts by name
}
\examples{
\dontrun{
zh_bezirke_suche("Affoltern")
zh_bezirke_suche("Affoltern", .include_query = TRUE)
}
}
