% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemeinden.R
\name{zh_gemeinde}
\alias{zh_gemeinde}
\title{Get a single municipality by code}
\usage{
zh_gemeinde(gemeinde_code)
}
\arguments{
\item{gemeinde_code}{Integer or character. The municipality code.}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble()}} with municipality details.
}
\description{
Get a single municipality by code
}
\examples{
\dontrun{
zh_gemeinde(261)  # Zürich
}
}
