% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemeindemutationen.R
\name{zh_gemeindemutationen}
\alias{zh_gemeindemutationen}
\title{List all municipality mutations (Gemeindemutationen)}
\usage{
zh_gemeindemutationen()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with municipality mutation information.
}
\description{
Retrieves all municipality mutation records (name changes, mergers, etc.)
in the Canton of Zurich.
}
\examples{
\dontrun{
zh_gemeindemutationen()
}
}
