% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemeinden.R
\name{zh_gemeinden}
\alias{zh_gemeinden}
\title{List all municipalities (Gemeinden)}
\usage{
zh_gemeinden()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\code{gebietstyp_code}, \code{gemeinde_code}, \code{gemeinde_name}.
}
\description{
Retrieves all municipalities in the Canton of Zurich.
}
\examples{
\dontrun{
zh_gemeinden()
}
}
