% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemeinden.R
\name{zh_gemeinden_suche}
\alias{zh_gemeinden_suche}
\title{Search municipalities by name}
\usage{
zh_gemeinden_suche(gemeindename, .include_query = FALSE)
}
\arguments{
\item{gemeindename}{Character scalar. (Partial) name to search for.}

\item{.include_query}{Logical. If \code{TRUE}, a \code{query} column with the
search term is added. Default \code{FALSE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{gemeinde_code},
\code{gemeinde_name}, \code{bezirk_code}, \code{bezirk_name},
\code{raumplanungsregion_code}, \code{raumplanungsregion_name}
(and optionally \code{query}).
}
\description{
Search municipalities by name
}
\examples{
\dontrun{
zh_gemeinden_suche("Tha")
zh_gemeinden_suche("Tha", .include_query = TRUE)
}
}
