% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemeindenhist.R
\name{zh_gemeindenhist}
\alias{zh_gemeindenhist}
\title{List all historical municipalities (Gemeindenhist)}
\usage{
zh_gemeindenhist()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with historical municipality information.
}
\description{
Retrieves all historical municipality records across all years in the
Canton of Zurich.
}
\examples{
\dontrun{
zh_gemeindenhist()
}
}
