% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{zh_health}
\alias{zh_health}
\title{Check API health}
\usage{
zh_health()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with API health information.
}
\description{
Calls the API health endpoint and returns the result as a tibble.
}
\examples{
\dontrun{
zh_health()
}
}
