% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raumplanungsregionen.R
\name{zh_raumplanungsregion}
\alias{zh_raumplanungsregion}
\title{Get a single spatial planning region by code}
\usage{
zh_raumplanungsregion(region_code)
}
\arguments{
\item{region_code}{Integer or character. The region code.}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble()}} with region details.
}
\description{
Get a single spatial planning region by code
}
\examples{
\dontrun{
zh_raumplanungsregion(101)  # Stadt Zürich
}
}
