% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raumplanungsregionen.R
\name{zh_raumplanungsregionen}
\alias{zh_raumplanungsregionen}
\title{List all spatial planning regions (Raumplanungsregionen)}
\usage{
zh_raumplanungsregionen()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
\code{gebietstyp_code}, \code{raumplanungsregion_code},
\code{raumplanungsregion_name}.
}
\description{
Retrieves all spatial planning regions in the Canton of Zurich.
}
\examples{
\dontrun{
zh_raumplanungsregionen()
}
}
