% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raumplanungsregionen.R
\name{zh_raumplanungsregionen_suche}
\alias{zh_raumplanungsregionen_suche}
\title{Search spatial planning regions by name}
\usage{
zh_raumplanungsregionen_suche(raumplanungsregionname, .include_query = FALSE)
}
\arguments{
\item{raumplanungsregionname}{Character scalar. (Partial) name to search for.}

\item{.include_query}{Logical. If \code{TRUE}, a \code{query} column with the
search term is added. Default \code{FALSE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{gemeinde_code},
\code{gemeinde_name}, \code{bezirk_code}, \code{bezirk_name},
\code{raumplanungsregion_code}, \code{raumplanungsregion_name}
(and optionally \code{query}).
}
\description{
Search spatial planning regions by name
}
\examples{
\dontrun{
zh_raumplanungsregionen_suche("Zürich")
zh_raumplanungsregionen_suche("Zürich", .include_query = TRUE)
}
}
