% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhgebiet-package.R
\docType{package}
\name{zhgebiet-package}
\alias{zhgebiet}
\alias{zhgebiet-package}
\title{zhgebiet: Client for the Gebietsstammdaten Kanton Zürich REST API}
\description{
Provides a tidyverse-friendly interface to the Gebietsstammdaten REST API
of the Canton of Zurich (\url{https://gebietsstammdaten.statistik.zh.ch}).
Allows users to retrieve information about municipalities (Gemeinden),
districts (Bezirke), spatial planning regions (Raumplanungsregionen),
and municipality assignments (Gemeindezuweisungen).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/metaodi/zhgebiet}
  \item Report bugs at \url{https://github.com/metaodi/zhgebiet/issues}
}
}
\author{
\strong{Maintainer}: metaodi \email{metaodi@users.noreply.github.com}
}
\keyword{internal}
